GetInstruction -> uint16_t

+++ innative/utility.cpp
@@ -136,7 +136,7 @@ namespace innative {
       return h;
     }
 
-    uint8_t GetInstruction(StringSpan ref)
+    uint16_t GetInstruction(StringSpan ref)
     {
       static const kh_opnames_t* h = GenOpNames();
